backend/answer_generation/requirements.txt
# RAG Answer Generation Dependencies
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0